Refcatored context integration tests.
Made the context tests use the same structure
as the structure in the lib.
