+++ tss-esapi/tests/context_tests/main.rs
@@ -1,0 +1,4 @@
+#[path = "../common/mod.rs"]
+mod common;
+mod general_esys_tr_tests;
+mod tpm_commands;
