+++ tss-esapi/tests/context_tests/tpm_commands/authenticated_countdown_timer_tests.rs
@@ -1,0 +1,2 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
