+++ tss-esapi/tests/context_tests/tpm_commands/dictionary_attack_functions_tests.rs
@@ -1,0 +1,2 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
