+++ tss-esapi/tests/context_tests/tpm_commands/duplication_commands_tests.rs
@@ -1,0 +1,2 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
