+++ tss-esapi/tests/context_tests/tpm_commands/mod.rs
@@ -1,0 +1,27 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
+mod asymmetric_primitives_tests;
+mod attached_components_tests;
+mod attestation_commands_tests;
+mod authenticated_countdown_timer_tests;
+mod capability_commands_tests;
+mod clocks_and_timers_tests;
+mod command_audit_tests;
+mod dictionary_attack_functions_tests;
+mod duplication_commands_tests;
+mod enhanced_authorization_ea_commands_tests;
+mod ephemeral_ec_keys_tests;
+mod field_upgrade_tests;
+mod hash_hmac_event_sequences_tests;
+mod hierarchy_commands_tests;
+mod integrity_collection_pcr_tests;
+mod miscellaneous_management_functions_tests;
+mod non_volatile_storage_tests;
+mod object_commands_tests;
+mod random_number_generator_tests;
+mod session_commands_tests;
+mod signing_and_signature_verification_tests;
+mod startup_tests;
+mod symmetric_primitives_tests;
+mod testing_tests;
+mod vendor_specific_tests;
